Rudamentary testing, small restructuring of dirs
To allow for testing, a test folder has been made. The pythonpath in
ones local environment might need to be updated to account for this. One
test implemented that ensures that our own version of a matrix
multiplication of 2x2 matrices give the correct output. This output has
been verified by hand.