Changed tests to pytest, added test cases for BDLOP
In order to get rid of a lot of print statements, an attempt has been
made to create a basic pytest file for BDLOP entries. This should check
for all things that we previously had spread out in print statements.
One test is currently failing, it is a known error.