Commitment scheme & BDLOP now use Cypari2, big refactor
A new class has been added for Polynomial utilities, simply called
"Polynomial". This uses the Cypari2 module as its base and allow the
user to get Gaussian or uniformly distributed elements in a ring. The
public facing method in this class lets the user specify a dimension and
get back a matrix or a vector of said elements. There are also some
helper functions that will come useful later.

This Polynomials class is now used by our commitment scheme as well as
by BDLOPZK. It has been tested and works for a q near 2^32, which the
previous implementation did not.

Because of these big changes, a lot of now unused files have been
removed.