Speed up of commitment scheme and BDLOP
When using Cypari, it accepts basically any input inside the Pari
object, as long as it is in a string. We had places where we would
concatenate strings by sending in f"{arr}, {arr2}~" to align them in
the proper dimension. This parsing takes incredibly long. Removing these
sped up the code so that 100 trials of the commitment scheme runs in
approximately two seconds.