Method in CommitmentScheme to output m, r, c for commit.
It was noted that oftentimes, users will not need to or want to set
their own polynomials for m and r which is required to commit. A method
now exists that generates these uniformly at random, commits them and
returns the Commit object (with m & r), as well as the c corresponding
to these. The tests had to be updated as well.