Added hash function to polynomial class, in use by BDLOP
In order to make BDLOP a NIZKP, we need to generate a hash of instance
specific parameters in the form of a polynomial. Because this polynomial
is often bounded by an l_1 norm of kappa, we have to supply kappa to the
method. It utilizes Numpy's Generator as well as PCG64 for pseudo-random
integers, while allowing for a seed. This ensures that the same
polynomial maps to the same hash which then maps with our implementation
to the same polynomial, always.