Cleanup of code base
Removed some files with print debugging to ensure basic functionality of
some of our core components. Also removed some files that were using
libraries that were no longer in use anywhere in the project and as such
has been deemed deprecated.

Moved classes and files to be in directories for their algorithm, so
that the commitment scheme we implemented is in use by BDLOP, as such it
is in the BDLOP folder (along with the actual BDLOP algorithm).