Made BDLOP into NIZKP, refactoring
Some of the previous refactoring made it impossible to generate a d in
a way that is required for a non-interactive zero-knowledge proof. Now d
is no longer sent as a parameter and it is up to the prover and verifier
to generate the same d. As such, d always relies on publicly available
data (not on r for example). Refactored and broke out common parts of
code into methods.