Move small invertible element from Comm_scheme to Polynomial
The "func_open" method in the commitmentscheme class was the difference
of two polynomials with an l_inf norm of 1 and a l_1 norm of kappa. This
function is needed in more places than in the commitment scheme, as such
it has been moved to the Polynomials class. The tests that were ran for
it has been moved as well.