Began work on DOTT21 as TDCommScheme, commit/open working
DOTT21 is the recommended Trapdoor commitment scheme. It uses the basis
of BDLOP which is already implemented in this project and then add a
trapdoor functionality. Got commit and open to work with the new
structure for A (in DOTT called Â, in BDLOP A1A2). Implemented some
basic testing to showcase this.