Update directories to use year of algorithm
There are times when conflicts can occur in Python when a directory
and a script share a name. It believes the script is a package or vice
versa and the program won't run on some machines. Updated so all
directories use the year of the algorithm, so BDLOP is now BDLOP16,
whereas the script file just use "BDLOP", without the year.