Increase memory of cypari instance
The memory allocated for Cypari was not enough for some of our big
matrix multiplications. Since we only have one instance, we only need to
change it at one point, in the Polynomials class. Also removed the
requirement for N to be a power of two, as this generalizes the class to
be usable for (t,n) threshold signatures.