Update one faulty test in secret_sharing, work on DOTT
One test was running with the wrong bounds in secret_share, causing it
to sometimes fail. That has been updated in this commit.

Additionally, work into the DOTT algorithm, with tests to accompany it.
There are some variables that are still unknown, all of these call to a
private method sigma, which right now returns a 0 for all cases.
Trapdoor commitment key generation now works, with tests to verify.