Gaussian element in polynomial class public, refactoring
There was some ugly code repetitions in the Polynomial class. These have
been refactored into one method that returns a lambda-expression for all
our cases. Also made it possible for other classes to fetch a single
gaussian element if needed. This will do the same as gaussian_array(1)
but allow for flexibility.