Found values of s, s_bar, added check in opening.
The values of Theta is as Max articulated to me just a different version
of big O notation. It means that the rate of growth should be exactly
equal to the specified value. Set s and s_bar to then be exactly as
articulated in the DOTT21 paper. Removed the theta function as it does
not do anything, and made B an inner function to the constructor.