Refactored Polynomials class, allow for bounded gaussian
The polynomials class had a lot of calls to functions that would make it
hard to scale, if one is to add a third distribution. Changed so there
is a private helper class that only needs to take in a method to
generate a single element from this distribution, and then returns an
element, a vector, or a matrix depending on the shape of the input.

Additionally there are a lot ot places where we will generate a bounded
Gaussian array, implemented so the Polynomials class can do that now.