Began working on big BGV refactoring of participants
The new BGV class, located under the BGV122 dir, is intended to let
attributes be confined to a certain instance of a class. We have a
Participant class that has methods for generating their own set of
secret values and commitments, and they then share this via the BGV
class. There are some generic methods that allow sharing of data between
all t parties, where an attribute in the class is set as a tuple of all
entries that are not what the instance contributed. This way, if we have
t=4, all instances should have other_a of length 3, all different of
course. Work done up until step 3 of the algorithm, not done with step 3
entirely.