Secret sharing in Keygen BGV, creating b_bar_i_j
Secret sharing now works. Originally each party computes their own
secret shares and share these with the others. As we have introduced a
"server" that is trustworthy (the BGV class itself), we send the e_i and
s_i to the server and let it do the sharing. This so that the
Participant can only worry about itself and not have to think about
states of other Participants.