Broadcast of all steps, valid openings for all check added
We now broadcast all commitments to each secret shared part, and
afterwards we broadcast the Commit object and each party ensures that
these open up legitimately. Because we always want to send in a function
of an honest_func, some refactoring was done in CommitmentScheme, namely
that this function is optional. If none is sent in, we send in a
polynomial "1" object as this will not affect any calculations.