Cleanup of code base
The other day it was brought to my attention that all of our
calculations wrapped in a Cypari object were superfluous. The Cypari
module is good enough to not require it, and can handle any arithmetic
on matrices and vectors without it. Removed the fixture that gets a
cypari instance, as none of the tests needed it anymore.

In addition, added a test to show how the additive homomorphic property
holds for our secret sharing scheme.