BGV rewrite, hash check for a, b, poly check for b_bar working
Due to some misunderstandings in the BGV Protocol scheme, I had shared
too much information between the participants. Rather than sharing s_bar
and e_bar, we share b_bar and since the secret sharing is homomorphic,
we can recreate it the same way. Required some major rewrites to the BGV
class and the Participant class. Throws an error if any hash breaks, or
if the secret sharing does not produce the same polynomial.