BGV Keygen rewrite working.
The current rewrite of BGV is now working for key generation. All
parties ensure checks that their commitments match the c they received,
as well as attempting to reconstruct their polynomials. To be
implemented is still a Proof of Relations. Right now if any of the
criteria fail, the program will fail at other steps further along.