Rewrite of BGV finally working!
Due to collaborating with Max, we managed to track down the source of
our errors and find out what was causing the issues, making the rewrite
of BGV functional at last. We still are missing commitments to the
randomness in the encryption and checking for it, as well as relation
proofs in the key-gen phase, but the main functionality exists and
works.