Refactoring of BGV rewrite to accept any q, p.
There were some hard-coded variables in comb that relied on our specific
choice of q, p. These have been changed to now instead be calculated
based on what we have set q and p to. In addition, now the secret keys
are returned in a list, and t_dec requires a list of secret keys. This
was done to get better type hints when supplying arguments for the
functions. Made a type alias for cypari2.gen.Gen to be a "poly".