Update Polynomial with new type, added RNG
Replaced all occurrences of "cypari2.gen.Gen" with new "poly" type. Also
made the Polynomial class have a random number generator, rather than
rely on numpy's built-in randomness. Currently the one we are
initializing is the same as the built-in one, but now it is easier to
change it in the future.

Removed TODO of verifying that our Gaussian elements are truly a
discrete Gaussian. They are not. We write explicitly that the
distribution is a rounded continuous Gaussian distribution.