BGV restructured to inherit from Controller superclass
When working with GKS23, it is immediately apparent that it will require
a lot of similar methods that have to do with sharing of data between
participants. Rather than duplicating code, I opted to abstract this
into a Controller class that both will inherit from. Will need to do the
same for participants in a later commit.