Ciphertext in BGV now its own object
Rather than keeping track of which u and v pair belong to one another in
GKS, it is easier if we make these into an object and keep the data
coupled. A Ctx object has been added that consists of a u and v.
Addition for ciphertexts is allowed, adding the corresponding u and
corresponding v parts.