Updated BGV, participants now handle enc/tdec/comb
A quick approach to BGV was to let the controller (the BGV class itself)
handle encrypting, decrypting and combining shares. This is not in-line
with how the GKS23 article describes their procedure, and proved to be
problematic as we need that functionality for participants in GKS23. Now
we do not return a pk/sk pair and instead leave these with the
participants themselves. Methods to call enc, t_dec and comb still exist
in BGV, these simply call the methods in a participant object that has
to be passed along.