GKS Rewrite passive security, sign and verify functional
There was a problem with the size of the default values I had set for
GKS23. A higher q led to the values not falling out of bounds and
producing the same errors. Now we also compare the hashes as it is
supposed to be done, returning true if and only if they are the same.