Active secure GKS23. Verifying Opening, hash of commit
The changes in GKS23 to make it into an actively secure protocol
involves hashing the sum of all commitments, no longer sending w, and
verifying opens for all other users. All of this has been implemented.
The check now only succeeds approximately 75% of the time when tested
multiple times. It fails due to one value in the second column of c
being too large, specifically in the position x^1023.