BGV now uses ternary distribution rather than Gaussian
BGV was using a Gaussian distribution for its s, e with a low sigma. In
the revision of GKS23 it is obvious that these are from a ternary
distribution of {-1, 0, 1} for all elements. The kappa chosen is 39 to
match Dilithium specs for 128 bit security, a comment is made to reflect
that.

In addition, I updated the new value of q to be in a power of 2 instead
of just the value.