Refactoring participants for BGV, GKS and interface
The rework yesterday introduced some changes that have to be specific to
their instance of the protocol running. BGV need its own comm scheme for
example, to handle the ciphertexts properly. Modified so that we do not
get superfluous parameters in protocols not used, BGV only needs its own
comm scheme, it does not need to know of its parents comm_scheme.

In addition, modifications to always send in a tuple of values for tn,
rather than two ints.