Update Kappa in GKS23 and BGV12.
Kappa should be a standalone variable in GKS23 and BGV12. in GKS23 the
commitment scheme's Kappa was used at all times when necessary, which
might cause problems if one has to change in the future and not the
other. In BGV12, both the comm scheme's Kappa was used, as well as one
specific to BGV12. This has been changed, now Kappa is consistently 36
across all schemes requiring it.