Remove print statements for cleanup
As the end of the thesis comes closer, we thought to remove some print
statements and the main method from GKS23. This was done so that anyone
can use our schemes without getting their terminals filled with
execution time statistics. These are still available on a remote branch
for anyone who wants to verify results and perhaps compute their own
time analysis.